QDoc: Use clang's C++ API to recognize copy/move constructors
QDoc uses clang's C API, LibClang, to parse the source code of C++
projects, so as to be able to extract the user-written documentation and
provide certain guarantees about its correctness.

When QDoc encounters a copy or move constructors, it marks its internal
representation of the element with this information, which is later used
to organize, order and render the documentation for the element.

As LibClang does not provide first party support for extracting
this information, a manual way to retrieve the information based on the
spelling of the argument of the processed constructors was performed.

This implementation was historically bugged.

As it did not follow the specification for copy/move constructors it was
able to recognize many false positives and unable to recognize certain
copy/move constructors.

Recent changes have solved some of the bugs that were acting as blocker
for other features but the implementation was left bugged as a
by-specification recognizer would have required a relative huge amount
of bug-prone work.

To solve the issue a series of patches were pushed upstream to LLVM,
exposing the required functionality of the C++ API to LibClang, which
was expected to be available with version 16, with the intention of
holding a fix until the upgrade to that new version was available.

Recent changes to QDoc allows us to turn to clang's C++ API, which is more
complete, on the spot from a LibClang's cursor thus not requiring any
more waiting time to solve the bugged implementation.

Hence, the detection method for copy/move constructors was revamped in
term of the C++ API, resolving the above mentioned issues.

`ClangCodeParser::processFunction` is the method that generally takes
care of extracting this form of information, enriching a `FunctionNode`,
the internal representation for documentable elements that are callable,
with the retrieved information.

`ClangCodeParser::processFunction` would perform ad-hoc checks based on
the spelling of the arguments of a processed cursor that was a
constructor.

The ad-hoc checks are thus removed, in favor of calling the
`CXXConstructorDecl::isCopyConstructor` and
`CXXConstructorDecl::isMoveConstructor` methods.

Certain constructors in the generated documentation may now be
re-ordered or marked differently in the output documentation if they
were incorrectly recognized or not recognized as copy/move constructors.

Change-Id: Id4e88cfa682b13899dd42af9c952f1316849dfee
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>