Qt Designer/calculatorbuilder example: Remove the form class and its autoconnected slots
The autoconnection feature is considered error-prone and is warned
about by clazy. As this removes the need for a class, remove it entirely
and use a lambda for the slot.

As a drive-by, fix the documentation of the project files.

Task-number: QTBUG-110447
Change-Id: Idf78b58445a3f89399f7ae09065f145428ef83fe
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>
(cherry picked from commit 0e5a54f96dda1ea72fdc43026e3d9bb74886c61e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>