HelpEngineWrapper: fix API susceptible to GCC 13 -Wdangling-reference
GCC 13 currently assumes that the temporary argument (the default
value) ends up embedded in the result, and warns:

  helpviewer.cpp:393:34: error: possibly dangling reference to a temporary [-Werror=dangling-reference]
    393 |         const HelpEngineWrapper &helpEngine = HelpEngineWrapper::instance();
        |                                  ^~~~~~~~~~
  helpviewer.cpp:393:74: note: the temporary was destroyed at the end of the full expression ‘Qt6::HelpEngineWrapper::instance(Qt6::QString())’
    393 |         const HelpEngineWrapper &helpEngine = HelpEngineWrapper::instance();
        |                                               ~~~~~~~~~~~~~~~~~~~~~~~~~~~^~

An attempt to fix the issue by passing the QString by value made it
worse, as all calls threw the warning then, including those that pass
a QString instead of relying on the defaulted argument.

So, heed QTBUG-98117 and overload instead of using default
arguments. Having the nullary overload in the same TU as the unary one
seems to make GCC shut up about the False Positive dangling.

Task-number: QTBUG-98117
Pick-to: 6.5 6.4 6.2 5.15
Change-Id: I69da70dfbf996ab923078631031f6f3c7a5cfcbe
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>