Doc: QDoc Manual: correctly output /*! and */ comment delimiters
The QDoc manual has a lot of code snippets demonstrating the use of
commands, and these comments include the C-style comment delimiters,
/*! and */. However, because */ is forbidden to appear within a comment
block (as it terminates it), the snippets included malformed versions
of the delimiters. This made copy-pasting annoying as the snippets
didn't represent valid QDoc syntax.

Fix the issue by using a '*' argument passed to code quoting commands,
allowing /\1! and \1/ to expand to the actual delimiters in the output.

Also,
 * Replace the use of \code with \badcode for QDoc snippets, avoiding
   the needless parsing/highlighting of non-C++ code.
 * Fix other minor issues in passing.

Pick-to: 6.5
Fixes: QTBUG-110142
Change-Id: Idcc4dc0fdfc6a8c7d368f3b2bc8f0dd8f1eb0843
Reviewed-by: Andreas Eliasson <andreas.eliasson@qt.io>
Reviewed-by: Venugopal Shivashankar <Venugopal.Shivashankar@qt.io>