QtResourceEditorDialog: Disconnect from QtQrcManager signals in d'tor
When QtResourceEditorDialog destructor is being run, the following
happens in exactly this order:
1. Explicit QtResourceEditorDialog d'tor is being run.
2. Scoped d_ptr of QtResourceEditorDialogPrivate goes out of scope,
   so d'tor of d_ptr is being run
3. Children of QtResourceEditorDialog are being deleted
4. One child of QtResourceEditorDialog was a QtQrcManager, and its
   d'tor is sending signals. This signals are still connected to
   QtResourceEditorDialog and respective lambda slots are called
   causing crash since all lambdas refer to d_ptr that was already
   deleted.

The fix is to disconnect from QtQrcManager signals in d'tor
so that it prevents calling lambdas referring to d_ptr.

The regression went in when replacing private slots with lambdas.
Apparently, when having private slots, the connections were
broken on destruction of d_ptr, while with lambdas it's not the
case anymore.

Amends 3dcf1490cb2bce34cb1f7777dfe836310f305817

Pick-to: 6.5
Fixes: QTBUG-110881
Change-Id: If42d0014d8bff9aae09439fcc453dd2c1e287ae6
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>