Designer: Don't leak menu bar in Docked Window mode
It looks like in Docked Window mode the main menu bar
is parentless (on platforms other than Q_OS_MACOS), while
in Multiple Top-Level Windows mode it has a parent.
In first case the menu bar was leaked, while in the second
one it was deleted by parent before running QDesignerWorkbench d'tor.

The fix is to keep a QPointer to menu bar instead of raw pointer
and delete it always from QDesignerWorkbench d'tor. When the
menu bar was already deleted, the QPointer to menu bar is cleared and
delete m_globalMenuBar() is no-op.

Detected by Valgrind Memory Analyzer.

Pick-to: 6.5
Change-Id: I0b3a77ec350a52e4b1eda532dd2a98236ec57bd6
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>