Qt Designer: Fix palette mask handling to consider single groups
Whereas the palette resolve mask in Qt 5 had one bit for all 3 colors
of a role, the mask is now a 64bit value with one bit for each
combination of role/group. Introduce helper functions (copied from
the QPalette source) to get the correct mask bits for each role/group
and for all groups of a role.

In the multiselection property helpers, check each role/group.

In the palette editor, use row masks to clear rows.

Pick-to: 6.5 6.4
Task-number: QTBUG-110963
Change-Id: I089a75d03828ff894527601fc111d5e39183e695
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>