qdoc: Improve Q_PROPERTY parsing for the property type
QDoc had a naive strategy for parsing Q_PROPERTY() macro arguments for
resolving the type and name; when the argument string is separated by
spaces, the first item was assumed to be the type, and the second item
the name. This failed for following valid declarations:

  Q_PROPERTY(QMap<QString, QString> map ...) // type contains a space
  Q_PROPERTY(Type * handle ...) // '*' interpreted as the name

Fix this by locating the first property attribute and then splitting
the input at that boundary. This allows QDoc to extract the property
type and name without making any assumptions about their syntax.

Remove old workarounds as unnecessary and implement safeguards for
invalid input. Also handle the MEMBER attribute correctly as it
implies the property is writable, whether the WRITE accessor is set
or not.

Fixes: QTBUG-111093
Change-Id: Ic1bf071a7b003a1cd0cb9a212dabdb7ea00ebfbe
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Luca Di Sera <luca.disera@qt.io>
(cherry picked from commit 6e68db8c10001f58fcc1d93e2e7291ac3ac7d42d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>