QDoc: Use clang's C++ API to extract parameters' types from a function
QDoc uses clang's C API, LibClang, to parse the source code of C++
projects, so as to be able to extract the user-written documentation and
provide certain guarantees about its correctness.

When QDoc encounters a function, method or similar element, it extracts
the types of the parameters of the element, both to show them to the
user and to compare them to the one provided by the user in a certain
document-block, to ensure consistency.

LibClang has a good support for those kind of extraction, such that they
are used to perform this effort in `ClangCodeParser::processFunction`,
the method that generally execute this process to enrich a
`FunctionNode`, the internal representation for documentable elements
that are callable, with the relevant information.

Recent changes to QDoc allows us to interleave usages of Clang's C and
C++ API, something that was done for certain parts of
`ClangCodeParser::processFunction` that could not be easily represented
with the more limited LibClang.

While there is no particular issue with the interleaving of the two
APIs, the two axis on which information is retrieved tend to require
certain overlapping operations, which lengthen the code and tend to not
mesh well with one another.

Thus, we refactor a further part of `ClangCodeParser::processFunction`,
to slowly convert it completely to a Clang's C++ API based method.

Hence, the code that takes care of extracting information about the
type of the parameters of a callable element was rewritten using the
equivalent Clang's C++ API calls.

Change-Id: I365cd1257b2a6a07187dbf23a971244b4c1f2daf
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>