qdoc: Add \details command
The command \details generates a collapsible <details>
element with a <summary> to control the hidden/visible
state.

The idea behind this is to provide more information with
less text on our main pages.

For output formats other than HTML, the details are
rendered as a normal paragraph and the summary string
is ignored.

Pick-to: 6.5
Change-Id: Id5daddd4854f22b9ceaa9a55b98af94a1ba06039
Reviewed-by: Paul Wicking <paul.wicking@qt.io>