qtattributionsscanner: Add SecurityCritical attribute
Serves as an indicator in the release process that these components need
to be carefully monitored and updated (even more often than the other
third-party modules).

So far this is not reflected in the generated documentation. This might
change in the future though.

For reasoning, see also

https://lists.qt-project.org/pipermail/development/2023-February/043667.html

Change-Id: I82c59e0198fc2fdc855aed89aa49f929391aa0ef
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 162573d4274e703cea7d95e4ea16158b3bbf710f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>