qdoc: Clean up usage of QDocDatabase::findAll* functions
QDocDatabase has a set of functions to traverse through all trees in
the forest and collect nodes that match specific criteria into static
maps for later retrieval. These functions may (and are) called multiple
times, so QDoc tries to ensure that each tree/function pair is executed
only once, and a reference to an existing map is returned on subsequent
calls.

Instead of storing a boolean in a Tree for marking it 'analyzed' and
having convoluted conditionals for re-executing a search function,
maintain a hash of functions per tree that are already processed. This
makes the code more readable and maintainable.

Change-Id: I6d610e7f86c9e250f0b050fd35f90bdf5d04396b
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Luca Di Sera <luca.disera@qt.io>