qdoc: Display individual enum items in \sincelist output
The output of \sincelist command lists enumerations introduced in a
specific version, documented with \enum with an associated \since
meta-command. Individual enum \value entries can also be marked with
a since-clause, however, this information was missing from the
generated list.

Adding support for this is less than trivial because EnumItem is not
a Node subclass, and the Sections class, which is responsible for
distributing the entries displayed in a since-list, operates on
Node instances. To work around this,

  * Introduce since()/setSince() to EnumItem
  * Resolve since info for enum values in Tree::resolveSince()
  * Add enum values to a dedicated map in QDocDatabase
  * Add enum value since attributes to .index file
  * Add 'New Enum Values' as a new Section entry

In order to generate content for the new section, don't directly
populate it; only add a single entry to mark it non-empty. Actual
enum values are then fetched from the new 'since-map'. Entries in
this map contain a pointer to the associated EnumNode, used for
linking.

Add a simple test case in generatedoutput autotest.

Fixes: QTBUG-110781
Change-Id: I064446dc1e369013284ba934adeabe908754b3bb
Reviewed-by: Luca Di Sera <luca.disera@qt.io>