QDoc: flatten loop in PureDocParser::processQdocComments
TLDR; Inverse the conditional to continue the while early, in order
to flatten the rest of the method.

PureDocParser::processQdocComments consists of a while loop over
a token. For any token that isn't Tok_Doc, get the next token and
continue. Tok_Doc tokens are processed, and this processing
constitutes the bulk of the method.

Inverse the logic of the conditional to allow an early continue. This
allows for dropping the else branch entirely, and makes it clear early
on that the method is only interested in Toc_Doc. Another advantage is
that the bulk of the method can be dedented to the main scope of the
while loop, reducing the cognitive load for readers of this code.

Change-Id: I01c9d3f547cd7610aa5adc19ffac456096f0a427
Reviewed-by: Luca Di Sera <luca.disera@qt.io>