qdoc: Add meta description header to HTML output
A meta description is an HTML element that provides a brief summary of
a web page, typically displayed as part of a search engine results
page.

Use the \brief description for pages that have it. This should cover
virtually all of the API reference as type \brief appears on module
pages and therefore is rarely missing. Many overview \page's have it
as well.

Pick-to: 5.15 6.2 6.4 6.5
Fixes: QTBUG-111360
Change-Id: Ie403777fd895a995bbd7d717333e95c1b6b77d89
Reviewed-by: Kai Köhne <kai.koehne@qt.io>