QDoc: Remove `PureCodeParser::m_token`
`PureCodeParser`, a class whose responsibility is to extract
documentation from non-programming-language files (e.g .qdoc files),
uses a `Tokenizer` to lex its processed source files.

During this processing, it uses an instance member, `m_token`, to store
the kind of the next token in the source input.

Nonetheless, the processing of a source file is well-scoped to a single
function, `PureCodeParser::processQdocComments`, the only user of
`m_token`, so that `m_token` does not require the instance-scoping that
it has.

Hence, `m_token` was removed in favor of a smaller scope variable in
`PureCodeParser::processQdocComments`.

The initialization of `m_token`, that is, the extraction of the first
token from an input file, was previously executed in
`PureDocParser::parseSourceFile`, the entry point for
`PureDocParser::processQdocComments`, and has now been moved to
`PureDocParser::processQdocComments`, in a `token` variable that is
created in the scope of the method.

The usages of `m_token`, all local to
`PureDocParser::processQdocComments`, were modified to be usages of
`token`.

And hence the `m_token` member was removed as it is now dead-code.

Change-Id: Ifafd6fdef1ce86865431581063da8ceb2e05d3c7
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>