QDoc: Remove `PureDocParser::m_tokenizer`
`PureCodeParser`, a class whose responsibility is to extract
documentation from non-programming-language files (e.g .qdoc files),
uses a `Tokenizer` instance to lex its processed source files.

A reference to the `Tokenizer` instance is stored in a member variable,
`m_tokenizer`, that is later accessed when the capabilities of the
`Tokenizer` are needed.

The `Tokenizer` instance that is referenced by `m_tokenizer` is created
on the stack every time `PureDocParser::parseSourceFile`, the entry
point to the parsing of an input file, is called.

As the instance will be destroyed after the call to
`PureDocParser::parseSourceFile` ends, `m_tokenizer` will be left
pointing to an invalid memory location, ensuing undefined behavior
should it be accessed outside of the scope of a call to
`PureDocParser::parseSourceFile`.

Since no access to `m_tokenizer` is performed outside this scope, no bug
was produced from it.

`m_tokenizer` only usage is in `PureDocParser::parseQdocComments`, the
method that actually performs the parsing and whose only caller is
`PureDocParser::parseSourceFile`, so that the instance-scope is
unrequired.

Hence, `m_tokenizer` was removed in favor of a smaller scope variable in
`PureCodeParser::processQdocComments`.

The initialization of `m_tokenizer` that was previously executed in
`PureDocParser::parseSourceFile` is now moved to
`PureDocParser::processQdocComments`, in a `tokenizer` variable that is
created in the scope of the method.

The usages of `m_tokenizer`, all local to
`PureDocParser::processQdocComments`, were modified to be usages of
`tokenizer`.

The usages are further modified to use the member of object access operator
instead of the member of pointer access operator due to the now
non-pointer type of `tokenizer`.

Thus, `m_tokenizer` was removed as dead-code.

Change-Id: I5c9e092cc5fd3b67ada8433b88fec2c42e4f6975
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>