QDoc: Move `CodeParser::precompileHeaders` to `ClangCodeParser`
QDoc parses a series of source file in different languages and formats
to find the user-provided documentation.

The base parsing interface for elements that perform this process is
given by the `CodeParser` class, with the concrete implementation of the
process for different languages/formats give by its child classes.

`CodeParser`, as part of its interface, provides certain methods that
are only meaningful when processing specific programming languages
source.

For example, it exposes `CodeParser::precompileHeaders` whose purpose it
to provide a PCH for parsing C++ source files.

The method is only implemented and used in the `ClangCodeParser` child
class, the class that actually implements the processing of C++ source
files.

Hence, the method is now removed from `CodeParser`'s interface and
directly exposed as part of the interface of `ClangCodeParser`, to
reduce the surface of the method and to reduce the dependencies between
`ClangCodeParser` and the `CodeParser` interface, which is generally
expected to be removed in the future.

Change-Id: I243b81974c741589e7f644e1996cdb274e20502b
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>