QDoc: Remove UsingClause struct and its use
The documentation for UsingClause states that it's "supposed to
describe a using clause" and that "I" (here, Martin Smith, the one
that documented the struct and maintainer of QDoc at the time) believe
the struct not to be in use at all.

Since 816b967374aa714139ddc59676dd32702bed9f49 was introduced in 2017,
type aliases ('using clause') have been handled by clang in QDoc through
the \typealias command. However, the code that was responsible for
handling using clauses prior to clang becoming the C++ parser for QDoc
was never really cleaned up, leaving code that was read but never
written to.  This caused QDoc to create a QList<UsingClause> of size 0
in Tree::resolveUsingClauses() more than 20000 times while generating
the documentation for Qt.

As the only use that was left is essentially a noop, remove the
UsingClause struct along with its only user, Tree::resolveUsingClauses.

Task-number: QTBUG-58158
Fixes: QTBUG-111820
Change-Id: Id83d12d4c0c53ef8b096b892c71bca97b7fa6752
Reviewed-by: Luca Di Sera <luca.disera@qt.io>