QDoc: Remove `CppCodeParser::ExtraFuncData`
`CppCodeParser`, the class that generally takes care of providing the
semantics of topics and meta commands for QDoc, exposed an inner class
`ExtraFuncData`, that bundled together some of the data that QDoc might
use when processing certain elements of the documentation.

`ExtraFuncData` was instantiated once and never used.
Supposedly, it was the either the result of a never finished
refactoring or was suppressed away during a refactoring, without the
complete code being cleaned of it.

Nonetheless, it is now removed as dead code.

Its single unused instantiation in `CppCodeParser::processTopicCommand`
was removed as a consequence of the removal of the class.

Change-Id: I7212cd0538421bb3c5e60517f53afb2eb4bd90bf
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>