qdoc: C++ marker: Allow single quote as a separator in integer literals
Since C++14, single quotes can appear in integer literals as separators
to improve readability.

When quoting a file that contains such a literal, QDoc interpreted the
single quote as a delimiter for a char, and kept reading ahead for the
closing quote. This interfered with correct marking up of C++ code and
ultimately caused a warning:

(qdoc) warning: Something is wrong with qdoc's handling of marked code

as the plain code and marked-up code end up having different number of
lines.

Pick-to: 6.5
Fixes: QTBUG-111973
Change-Id: Icd2d95973f9cee72eb1c9e3f2d79df0af643dbaa
Reviewed-by: Paul Wicking <paul.wicking@qt.io>